+++ src/Services/ISwaggerService.cs
@@ -1,3 +1,6 @@
+using Microsoft.AspNetCore.Builder;
+using System.Collections.Generic;
+
 namespace PipServices3.Rpc.Services
 {
     /// <summary>
@@ -9,5 +12,12 @@ namespace PipServices3.Rpc.Services
         /// Perform required Swagger registration steps.
         /// </summary>
         void RegisterOpenApiSpec(string baseRoute, string content);
+
+        /// <summary>
+        /// Configure application to use Swagger UI
+        /// </summary>
+        /// <param name="applicationBuilder">application builder</param>
+        /// <param name="routes">list of routes to swagger documents</param>
+        void ConfigureApplication(IApplicationBuilder applicationBuilder, List<string> routes);
     }
 }