+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for .NET Changelog
 
+## <a name="3.3.6"></a> 3.3.6 (2020-12-09)
+
+### Features
+* Removed Swashbuckle package
+* Using ISwaggerService to configuure Swagger UI
+
 ## <a name="3.3.5"></a> 3.3.5 (2020-12-05)
 
 ### Features